requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
websocket-client==1.6.4
schedule==1.2.0
loguru==0.7.2
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0 